src/SFA.DAS.AANHub.Application/Admins/Commands/CreateAdminMember/CreateAdminMemberCommand.cs
using MediatR;
using SFA.DAS.AANHub.Application.Common;
using SFA.DAS.AANHub.Application.Mediatr.Responses;
using SFA.DAS.AANHub.Domain.Entities;
using static SFA.DAS.AANHub.Domain.Common.Constants;

namespace SFA.DAS.AANHub.Application.Admins.Commands.CreateAdminMember;

public class CreateAdminMemberCommand : CreateMemberCommandBase, IRequest<ValidatedResponse<CreateMemberCommandResponse>>
{
    public string UserName { get; set; } = null!;

    public static implicit operator Member(CreateAdminMemberCommand command) => new()
    {
        Id = command.Id,
        UserType = MembershipUserType.Admin,
        Status = MembershipStatus.Live,
        Email = command.Email!,
        FirstName = command.FirstName!,
        LastName = command.LastName!,
        JoinedDate = command.JoinedDate!.Value,
        RegionId = command.RegionId,
        OrganisationName = command.OrganisationName,
        Admin = new Admin
        {
            MemberId = command.Id,
            UserName = command.UserName,
        }
    };
}